lints: fix part of clippy warning

+++ src/daemon/src/runtime/migration/hostmem_buffer.rs
@@ -38,7 +38,7 @@ impl HostMemBufferManager {
 
     pub fn store(&self, buffer_id: &BufferId, data: &[u8]) -> Result<(), HybridBufferError> {
         let mut inner = self.inner.lock().unwrap();
-        if data.len() > MAX_ALLOCATION_SIZE as usize {
+        if data.len() > MAX_ALLOCATION_SIZE {
             return Err(HybridBufferError::InvalidInputBuffer);
         }
         if let Some(mut block_buffer) = inner.alloc_buffer() {
